requirements.txt
accelerate==0.25.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
asttokens==3.0.1
attrs==25.4.0
blinker==1.9.0
cachetools==7.0.2
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
datasets==2.16.1
debugpy==1.8.20
decorator==5.2.1
dill==0.3.7
executing==2.2.1
faiss-cpu==1.13.2
filelock==3.20.0
Flask==3.1.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2023.10.0
gitdb==4.0.12
GitPython==3.1.46
huggingface_hub==0.36.2
idna==3.11
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.15
narwhals==2.17.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.3.5
packaging==26.0
pandas==2.3.3
parso==0.8.6
pillow==12.0.0
platformdirs==4.9.2
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pure_eval==0.2.3
pyarrow==23.0.1
pyarrow-hotfix==0.7
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.14.3
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==2.6.1
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
streamlit==1.55.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.15.2
toml==0.10.2
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==4.37.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
wcwidth==0.6.0
Werkzeug==3.1.6
xxhash==3.6.0
yarl==1.23.0
gunicorn